On branch refractoring  Changes to be committed: 	modified:   README.md 	new file:   json/config.json 	new file:   json/rules.json 	modified:   src/main/java/ru/discordj/bot/config/MonitoringManager.java 	modified:   src/main/java/ru/discordj/bot/events/listener/configurator/command/MonitoringCommand.java 	renamed:    src/main/java/ru/discordj/bot/informer/ServerMonitor.java -> src/main/java/ru/discordj/bot/monitor/ServerMonitor.java 	renamed:    src/main/java/ru/discordj/bot/informer/parser/Arma3ServerQuery.java -> src/main/java/ru/discordj/bot/monitor/parser/Arma3ServerQuery.java 	renamed:    src/main/java/ru/discordj/bot/informer/parser/DayZServerQuery.java -> src/main/java/ru/discordj/bot/monitor/parser/DayZServerQuery.java 	renamed:    src/main/java/ru/discordj/bot/informer/parser/Parser.java -> src/main/java/ru/discordj/bot/monitor/parser/Parser.java 	renamed:    src/main/java/ru/discordj/bot/informer/parser/ServerQuery.java -> src/main/java/ru/discordj/bot/monitor/parser/ServerQuery.java 	renamed:    src/main/java/ru/discordj/bot/informer/parser/SteamInputStream.java -> src/main/java/ru/discordj/bot/monitor/parser/SteamInputStream.java 	renamed:    src/main/java/ru/discordj/bot/informer/query/GameServerQuery.java -> src/main/java/ru/discordj/bot/monitor/query/GameServerQuery.java 	renamed:    src/main/java/ru/discordj/bot/informer/query/GameServerQueryFactory.java -> src/main/java/ru/discordj/bot/monitor/query/GameServerQueryFactory.java 	renamed:    src/main/java/ru/discordj/bot/informer/query/GameSpyQuery.java -> src/main/java/ru/discordj/bot/monitor/query/GameSpyQuery.java 	renamed:    src/main/java/ru/discordj/bot/informer/query/SourceQuery.java -> src/main/java/ru/discordj/bot/monitor/query/SourceQuery.java 	renamed:    src/main/java/ru/discordj/bot/informer/query/UT3Query.java -> src/main/java/ru/discordj/bot/monitor/query/UT3Query.java 	renamed:    src/main/java/ru/discordj/bot/informer/sender/ISender.java -> src/main/java/ru/discordj/bot/monitor/sender/ISender.java 	renamed:    src/main/java/ru/discordj/bot/informer/sender/Sender.java -> src/main/java/ru/discordj/bot/monitor/sender/Sender.java 	renamed:    src/main/java/ru/discordj/bot/informer/tools/Tools.java -> src/main/java/ru/discordj/bot/monitor/tools/Tools.java 	modified:   src/main/java/ru/discordj/bot/utility/pojo/Root.java
