+++ json/config.json
@@ -1,0 +1,14 @@
+{
+  "token" : "empty",
+  "owner" : "empty",
+  "inviteLink" : "empty",
+  "roles" : [ {
+    "channelId" : "empty",
+    "roleId" : "empty",
+    "emojiId" : "empty"
+  } ],
+  "monitoringChannelId" : null,
+  "servers" : [ ],
+  "monitoringEnabled" : false,
+  "currentMonitor" : null
+}
