+++ json/rules.json
@@ -1,0 +1,6 @@
+{
+  "title" : "█▓▒░⡷⠂\uD835\uDE83\uD835\uDE91\uD835\uDE8E \uD835\uDE82\uD835\uDE9D\uD835\uDE8E\uD835\uDE8A\uD835\uDE95\uD835\uDE9D\uD835\uDE91 \uD835\uDE73\uD835\uDE9E\uD835\uDE8D\uD835\uDE8E⠐⢾░▒▓█",
+  "welcomeField" : "WELCOME",
+  "rulesField" : "️1️⃣ Все участники сервера имеют равные права...",
+  "footer" : "\uD83D\uDCE9 requested by @{author} {date}"
+}
